refactor: decompose Toast component
Fixes React Doctor's no-giant-component finding (and #354): extract
ToastContent + ToastButton render components and the merged-styles,
animations (wiggle/stack transform), swipe-props, and lifecycle
(app-state timers, height measurement) hooks into their own modules.
Verbatim moves, no behavior change; Toast drops from ~580 to ~250
lines.
